Moved RecordState into webapp/store/ui/record/state
