+++ webapp/loggedin/surveyViews/reducer.js
@@ -1,12 +1,11 @@
 import { combineReducers } from 'redux'
 
-import * as RecordState from './record/recordState'
+import { RecordReducer, RecordState } from '../../store/ui/record'
 import * as SurveyFormState from './surveyForm/surveyFormState'
 
-import record from '../../store/ui/record/reducer'
 import surveyForm from './surveyForm/reducer'
 
 export default combineReducers({
-  [RecordState.stateKey]: record,
+  [RecordState.stateKey]: RecordReducer,
   [SurveyFormState.stateKey]: surveyForm,
 })
