+++ webapp/loggedin/surveyViews/surveyForm/nodeDefs/components/types/nodeDefCode.js
@@ -15,11 +15,12 @@ import * as Node from '@core/record/node'
 import * as NodeRefData from '@core/record/nodeRefData'
 import * as NodeDefLayout from '@core/survey/nodeDefLayout'
 
+import { I18nState } from '@webapp/store/system'
 import { SurveyState } from '@webapp/store/survey'
-import * as RecordState from '../../../../record/recordState'
+import { RecordState } from '@webapp/store/ui/record'
+
 import NodeDefCodeCheckbox from './nodeDefCodeCheckbox'
 import NodeDefCodeDropdown from './nodeDefCodeDropdown'
-import { I18nState } from '@webapp/store/system'
 
 const NodeDefCode = (props) => {
   const {
