+++ webapp/loggedin/surveyViews/surveyForm/nodeDefs/components/useValidation.js
@@ -9,8 +9,7 @@ import * as RecordValidation from '@core/record/recordValidation'
 import * as Validation from '@core/validation/validation'
 
 import { useSurvey } from '@webapp/store/survey'
-
-import * as RecordState from '@webapp/loggedin/surveyViews/record/recordState'
+import { RecordState } from '@webapp/store/ui/record'
 
 export default (props) => {
   const { edit, node, nodeDef, nodes, parentNode } = props
