+++ webapp/store/ui/record/index.js
@@ -1,4 +1,5 @@
 import * as RecordActions from './actions'
 import RecordReducer from './reducer'
+import * as RecordState from './state'
 
-export { RecordActions, RecordReducer }
+export { RecordActions, RecordReducer, RecordState }
